+++ docs/changelog.md
@@ -11,6 +11,7 @@
 ### Changes
 * Replace custom docs-checking job with Kurtosis docs-checker orb
 * Use the devtools script for updating package versions
+* Updated the index.ts to export ContainerRunConfig + ContainerCreationConfig
 
 ### Fixes
 * Fixed Node engine dependency in Typescript library to be `>=14.17.0`